Add Random Stuff
The file adds some random functions which are never even called and some random variables which again are never used. 
Also, it takes all the things that are in main to a function named solve_my_code_please and calls that in the main. 
This also adds an array to the function main and sorts that array using merge sort.